+++ server/src/main/kotlin/suwayomi/tachidesk/manga/impl/extension/ExtensionsList.kt
@@ -132,6 +132,8 @@ object ExtensionsList {
                                     // Always update icon url and repo
                                     this[ExtensionTable.iconUrl] = foundExtension.iconUrl
                                     this[ExtensionTable.storeIndexUrl] = foundExtension.storeIndexUrl
+                                    this[ExtensionTable.apkUrl] = foundExtension.apkUrl
+                                    this[ExtensionTable.jarUrl] = foundExtension.jarUrl
 
                                     // add these because batch updates need matching columns
                                     this[ExtensionTable.hasUpdate] = extensionRecord[ExtensionTable.hasUpdate]
@@ -175,6 +177,7 @@ object ExtensionsList {
                                     this[ExtensionTable.lang] = foundExtension.lang
                                     this[ExtensionTable.contentWarning] = foundExtension.contentWarning.ordinal
                                     this[ExtensionTable.apkUrl] = foundExtension.apkUrl
+                                    this[ExtensionTable.jarUrl] = foundExtension.jarUrl
                                     this[ExtensionTable.iconUrl] = foundExtension.iconUrl
                                 }
                             }.toExecutable()
@@ -192,6 +195,7 @@ object ExtensionsList {
                         this[ExtensionTable.lang] = foundExtension.lang
                         this[ExtensionTable.contentWarning] = foundExtension.contentWarning.ordinal
                         this[ExtensionTable.apkUrl] = foundExtension.apkUrl
+                        this[ExtensionTable.jarUrl] = foundExtension.jarUrl
                         this[ExtensionTable.iconUrl] = foundExtension.iconUrl
                     }
                 }