+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased] (Preview)
 
 ### Added
+- (**Downloads**) Try to preserve downloaded files during a chapter list update for chapters with title and/or scanlator change
 - (**Logs**) Add IP location logging
 
 ### Changed
@@ -14,6 +15,36 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 - .
+
+## [v2.3.2243] - 2026-07-13
+
+### Added
+- (**Extension**) Add ability to install jar versions of extensions
+
+### Fixed
+- (**Extension**) Fix missing memo field in getPageList
+
+## [v2.3.2238] - 2026-07-05
+
+### Changed
+- (**Extension**) Support fully qualified extension class names
+- (**Manga/API**) Return partial responses for Manga And Chapters query
+
+### Fixed
+- (**Database/H2**) Use temporary file for H2 migration tools download
+
+## [v2.3.2232] - 2026-07-03
+
+### Fixed
+- (**Extension**) Update APK url on installed Extensions
+
+## [v2.3.2230] - 2026-07-01
+
+### Fixed
+- (**Extension/API**) Fix GQL handling of extensions without an extension store
+- (**Build/Bundler**) Fix build continuing if errors happen
+- (**Build/Bundler**) Remove jogamp from build script
+- (**Build/Bundler**) Fix chmod attempts on Windows Java
 
 ## [v2.3.2223] + [WebUI: v20260509.01] - 2026-06-30
 
@@ -478,8 +509,12 @@ Huge thanks to @martinek who pulled the most of the weight this release!
 
 <!-- SERVER LINKS -->
 
-[unreleased]: https://github.com/suwayomi/suwayomi-server/compare/v2.3.2223...HEAD
-[v2.3.2223]: https://github.com/suwayomi/suwayomi-server/compare/v2.1.2100...v2.3.2223
+[unreleased]: https://github.com/suwayomi/suwayomi-server/compare/v2.3.2243...HEAD
+[v2.3.2243]: https://github.com/suwayomi/suwayomi-server/compare/v2.3.2238...v2.3.2243
+[v2.3.2238]: https://github.com/suwayomi/suwayomi-server/compare/v2.3.2232...v2.3.2238
+[v2.3.2232]: https://github.com/suwayomi/suwayomi-server/compare/v2.3.2230...v2.3.2232
+[v2.3.2230]: https://github.com/suwayomi/suwayomi-server/compare/v2.3.2223...v2.3.2230
+[v2.3.2223]: https://github.com/suwayomi/suwayomi-server/compare/v2.2.2100...v2.3.2223
 [v2.2.2100]: https://github.com/suwayomi/suwayomi-server/compare/v2.1.1867...v2.2.2100
 [v2.1.1867]: https://github.com/suwayomi/suwayomi-server/compare/v2.0.1727...v2.1.1867
 [v2.0.1727]: https://github.com/suwayomi/suwayomi-server/compare/v1.1.1...v2.0.1727